Re-introduce `Cached<R>`
Now, instead of reusing the same `Bump` arena inside the `Cached<R>`, the vdom
maintains a set of `Bump` arenas and cached `Node`s that are the results of
rendering `Cached<R>`s. The entries in the set are immutable after insertion,
and invalidating a `Cached<R>` causes it to request a new `Bump` arena from the
vdom instead of resetting the old `Bump` that it allocated into last.

But how to clean up this set of cached immutable `Bump` arenas? We have a tiny,
specialized garbage collector for just these arenas. Because they are immutable,
their outgoing edges never change, so we compute and store the set of transitive
edges when inserting new arenas into the set. This means that the "mark" phase
of our collector is super quick and easy, since it doesn't need to recursively
mark things or maintain a mark stack. We garbage collect the arenas after each
render.

Additionally, there is a new `NodeKind::Cached` variant. This contains the id of
its results in the set of cached nodes and their `Bump` arenas. When diffing, if
we see that the old and new virtual dom point to the same cached subtree, then
we avoid diffing them. This gives us really nice speed ups, and makes the new
`Cached<R>` even better than the old version, even ignoring the bugs that the
old version had.

Fixes #31.